Gtk3: Disable the resource-manager-widget
There's a bug somewhere in the resource-widget-manager use in the settings tree, which I can't find at the moment. To avoid segfaults in the UI, I've disabled its use for now, until I'm comfortable I've found the bug.
This means reset-to-dialog-entry/reset-to-default/reset-to-vicerc don't work at the moment, but still better than the UI mysteriously crashing.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@35460 379a1393-f5fb-40a0-bcee-ef074d9b53f7