Fix format specifier warnings
This one caused me some concern due to using the preprocessor. Integer literals are treated as signed int, so I had to add a few 'U' suffixes to literals in the macros to avoid type promotion, and a few casts.
Casts are usually dubious, especially combined with the cpp, so the people who actually wrote this code should check it (A. Fachat, A. Boose), I think I got it right, but I can't guarantee it.
Also, depending on a single-letter var 'b' being available in a macro call without using it as an argument seems wrong. (See PARALLEL_DEBUG_SET_BUS_TYPE)



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@36449 379a1393-f5fb-40a0-bcee-ef074d9b53f7