Gtk3: Comment out lines that cause warnings due to updated Gtk/Glib
Looks like the latest msys2 gtk/glib updates caused our old in-tree vte code to produce a lot of warnings. It all has to with background(color), but since we can't just update our in-tree Windows-fixed vte (GPL issue), we'll have to do with this, hoping for no further breakage. Great.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@37556 379a1393-f5fb-40a0-bcee-ef074d9b53f7