Apply patch #252 from Roberto Muscedere.
In vdrive, when formatting or validating an 8050 or 8250 image, the header
sector isn't flagged as allocated. This results in vice hanging when creating
more than 81 files on the disk (this happens after file 80). The 81st file
tries to write the directory information to the header sector, but doesn't,
however the last directory sector creates a loop chain which points back to the
header. Ultimately vice hangs up in this infinite sector chaining.
This fix addresses the issue in the validate command (which is used by format)
to allocate this sector. I have confirmed the TDE of the 8x50 does allocate
this sector after a format. (second patch in the attached file)
For 8250 images specifically, once the first "side" of the disk is full (track
1-77), it ends up corrupting data as the number of available tracks isn't
reported correctly to the BAM allocation system. A patch is included here. (the
first patch in the file).


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@38900 379a1393-f5fb-40a0-bcee-ef074d9b53f7