Lie a bit about the LED colours of the IEEE dual floppy drives.
Even though the drive activity LEDs are red in reality, we make them green for
now. But at least we can see the error LED in the GUI, which is red and in the
same location as the activity LED for drive 0.
Having the colour wrong seems less bad than not seeing the error indicator.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@40716 379a1393-f5fb-40a0-bcee-ef074d9b53f7