Gtk3: Fix 'Keyset Joysticks' hotkey
Another side-effect of the goddamn runtime emu checks: the 'keyset-joystick-toggle' UI action was present in two structs, one for emus with controlport-swap, one for emus without controlport-swap. The menu item iterator always picked the first, so keyset-joystick-toggle's hotkey was registered for emus with controlport-swap, ie NOT xvic, xcbm2 or xpet.
Also fixed check button synchronization for the keyset-joystick-toggle item.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@41254 379a1393-f5fb-40a0-bcee-ef074d9b53f7